+++ frontend/next-env.d.ts
@@ -1,12 +1,5 @@
-import NextAuth from "next-auth";
+/// <reference types="next" />
+/// <reference types="next/image-types/global" />
 
-declare module "next-auth" {
-  interface Session {
-    user: {
-      name?: string;
-      email?: string;
-      image?: string;
-      token?: string; //add your custom token here
-    };
-  }
-}
+// NOTE: This file should not be edited
+// see https://nextjs.org/docs/app/building-your-application/configuring/typescript for more information.
