+++ frontend/src/types/next-auth.d.ts
@@ -1,0 +1,12 @@
+import NextAuth from "next-auth";
+
+declare module "next-auth" {
+  interface Session {
+    user: {
+      name?: string;
+      email?: string;
+      image?: string;
+      token?: string; // custom token
+    };
+  }
+}
